Move Concast into its own module.
